Function casper_contract::contract_api::storage::new_locked_contract

pub fn new_locked_contract(
    entry_points: EntryPoints, 
    named_keys: Option<NamedKeys>, 
    hash_name: Option<String>, 
    uref_name: Option<String>
) -> (ContractHash, ContractVersion)

Create a locked contract stored under a Key::Hash, which can never be upgraded. This is an irreversible decision; for a contract that can be upgraded use new_contract instead. if named_keys are provided, will apply them if hash_name is provided, puts contract hash in current context’s named keys under hash_name if uref_name is provided, puts access_uref in current context’s named keys under uref_name