Module casper_contract::contract_api::runtime

Functions for interacting with the current runtime.

Functions

blake2b

Returns a 32-byte BLAKE2b digest

call_contract

Calls the given stored contract, passing the given arguments to it.

call_versioned_contract

Invokes the specified entry_point_name of stored logic at a specific contract_package_hash address, for the most current version of a contract package by default or a specific contract_version if one is provided, and passing the provided runtime_args to it

get_blocktime

Returns the current BlockTime.

get_call_stack

Returns the call stack.

get_caller

Returns the caller of the current context, i.e. the AccountHash of the account which made the deploy request.

get_key

Returns the requested named Key from the current context.

get_named_arg

Returns given named argument passed to the host for the current module invocation.

get_phase

Returns the current Phase.

has_key

Returns true if name exists in the current context’s named keys.

is_valid_uref

Validates uref against named keys.

list_named_keys

Returns the named keys of the current context.

put_key

Stores the given Key under name in the current context’s named keys.

remove_key

Removes the Key stored under name in the current context’s named keys.

ret

Returns the given CLValue to the host, terminating the currently running module.

revert

Stops execution of a contract and reverts execution effects with a given ApiError.